almost done

+++ src/main/java/ru/girfanov/tm/service/TaskService.java
@@ -31,26 +31,26 @@ public final class TaskService extends AbstractService<Task> implements ITaskSer
     }
 
     @Override
-    public Collection<Task> findAllTasksByProjectId(@NotNull final String projectUuid) {
-        if(projectUuid.isEmpty()) { return null; }
-        return taskRepository.findAllTasksByProjectId(projectUuid);
+    public Collection<Task> findAllTasksByProjectId(@NotNull final String projectId, @NotNull final String userId) {
+        if(repository.findEntityById(userId) == null || projectId.isEmpty()) { return null; }
+        return taskRepository.findAllTasksByProjectId(projectId);
     }
 
     @Override
-    public void removeAllTasksByProjectId(@NotNull final String projectId) {
-        if(projectId.isEmpty()) { return; }
+    public void removeAllTasksByProjectId(@NotNull final String projectId, @NotNull final String userId) {
+        if(repository.findEntityById(userId) == null || projectId.isEmpty()) { return; }
         taskRepository.removeAllTasksByProjectId(projectId);
     }
 
     @Override
     public Collection<Task> findAllTasksByUserId(@NotNull final String userId) {
-        if(userId.isEmpty()) { return null; }
+        if(repository.findEntityById(userId) == null) { return null; }
         return taskRepository.findAllTasksByUserId(userId);
     }
 
     @Override
     public void removeAllTasksByUserId(@NotNull final String userId) {
-        if(userId.isEmpty()) { return; }
+        if(repository.findEntityById(userId) == null) { return; }
         taskRepository.removeAllTasksByUserId(userId);
     }
 }
